+++ src/app/date/relativeTime/util/relativeTime.interface.ts
@@ -1,0 +1,5 @@
+import { Locale } from '../../types/locale';
+
+export interface RelativeTime {
+  relativeTime(date: Date, locale: Locale): string;
+}