+++ src/app/date/types/locale.ts
@@ -1,0 +1,10 @@
+/**
+ * String locale.
+ *
+ * @name Locale
+ * @typedef {'en-US' | 'es-ES'} Locale
+ * @alias Locale
+ */
+export type Locale =
+  | 'en-US'
+  | 'es-ES';