+++ src/test/integration/date/relativeTime/own/index.spec.ts
@@ -1,0 +1,4 @@
+import { RelativeTimeOwn } from '../../../../../app/date/relativeTime/impl/relativeTimeOwn';
+import { RelativeTimeTest } from '../test';
+
+new RelativeTimeTest(new RelativeTimeOwn()).runTest();